Added text generation
If a file exists called <templateName>.txt in the same dir as
<templateName>.html, its variables will be compiled with the same context
as the <templateName>.html rendering and the output returned

If such a file does not exists, the text will be autogenerated from the
html using the html-to-text module.

If the option text: false is set, the generateText function will return
null text.